Reuse method to get labels and move routing sidecar utils to workload.go
